dat editor: moved palettegridwidget into its own package
