+++ hswidget/palettegridwidget/doc.go
@@ -1,0 +1,3 @@
+// Package palettegridwidget provides data for viewing
+// and editing 256-colors palettes
+package palettegridwidget
